speedy-breakout: Add headers for feedback expansion
This consists of a footprint for an arbitrary 2.54 mm connector, with
traces to another header which can be used to easily solder wires to
make arbitrary connections to the Speedy breakout through-holes
depending on the desired application. A 6 pin footprint was used which
should give enough for power + ground and up to a full 4 pin SPI
interface. Pin 6 is tied to ground.